+++ community/logicaldoc/trunk/logicaldoc-core/src/main/java/com/logicaldoc/core/communication/dao/EMailDAO.java
@@ -13,44 +13,44 @@ import com.logicaldoc.core.communication.EMail;
  */
 public interface EMailDAO {
 
-    /**
-     * This method persists an emailaccount object.
-     * 
-     * @param account EMailAccount which should be store.
-     * @return True if successfully stored in a database.
-     */
-    public boolean store(EMail email);
+	/**
+	 * This method persists an emailaccount object.
+	 * 
+	 * @param account EMailAccount which should be store.
+	 * @return True if successfully stored in a database.
+	 */
+	public boolean store(EMail email);
 
-    /**
-     * This method deletes an email.
-     * 
-     * @param emailId The email identifier which should be delete.
-     */
-    public boolean delete(long emailId);
+	/**
+	 * This method deletes an email.
+	 * 
+	 * @param emailId The email identifier which should be delete.
+	 */
+	public boolean delete(long emailId);
 
-    /**
-     * This method finds an email by its id.
-     */
-    public EMail findByPrimaryKey(long emailId);
+	/**
+	 * This method finds an email by its id.
+	 */
+	public EMail findByPrimaryKey(long emailId);
 
-    public Collection<EMail> findByUserName(String username);
+	public Collection<EMail> findByUserName(String username);
 
-    public Collection<EMail> findByUserName(String username, String folder);
+	public Collection<EMail> findByUserName(String username, String folder);
 
-    /**
-     * Loads all e-mails downloaded by the specified account
-     * 
-     * @param accountId
-     * @return
-     */
-    public Collection<EMail> findByAccountId(int accountId);
+	/**
+	 * Loads all e-mails downloaded by the specified account
+	 * 
+	 * @param accountId
+	 * @return
+	 */
+	public Collection<EMail> findByAccountId(long accountId);
 
-    /**
-     * Same as findByAccountId except for the fact that the returned collection
-     * contains e-mail identifiers only
-     * 
-     * @param accountId
-     * @return
-     */
-    public Collection<String> collectEmailIds(int accountId);
+	/**
+	 * Same as findByAccountId except for the fact that the returned collection
+	 * contains e-mail identifiers only
+	 * 
+	 * @param accountId
+	 * @return
+	 */
+	public Collection<String> collectEmailIds(long accountId);
 }
